Fixed a bug where the program would behave weird after redo and then a letter input
when a user inputs a letter or a string instead of an integer, C language automatically assigns that variable to its previous value.
That's why when the user would redo and then input a letter, the program would assign row1 = 22 and go into redo once again, behaving not as expected.
fixed it by assigning row1 to -1 so my  function would pickup the mistake